+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes
 
+#### Fixes
+
+- Fix inconsistent changes with JSONField ([#355](https://github.com/jazzband/django-auditlog/pull/355))
+
 ## 2.0.0 (2022-05-09)
 
 #### Improvements